Add files via upload
Project Description
This project is about End-to-end encryption (E2EE) which is a method of secure communication that prevents third parties from accessing data while it’s transferred from one end system or device to another. It is designed to be user friendly and includes 4 different modes to switch. The text entered will be encrypted in the user’s desired mode only when the secret code is correct.

Background Information
Encryption is the process of translating plaintext data into ciphertext (which is not in human readable form). Decryption is the process of converting ciphertext back to plaintext.

Motivation
These days, people are using variety of messaging platforms to communicate online. The larger their digital footprints become, the more private and secure communication tools are in need. The best solution to this problem is E2EE which can keep our data safe form hacks and unauthorized access.